change hash for symbols in constants to 64 bits to avoid collisions, be sure single-step is cleared on ad1 exit

+++ Tools/AM1/eval.c
@@ -347,14 +347,14 @@ SymNodeP symP;
             bigint += (0xEFE << 20);
         }
 
-        return( (int)bigint );
+        return( (long int)bigint );
 
     case WILDREF:
         // All we have is the symbolname, use the string address
         bigint = (uint64_t)(nodeP->value.strP);
         bigint = (bigint & 0xFFFFFFFF) << 18;  // mangle its bits 
         bigint += (0xEEF << 20);
-        return( (int)bigint );
+        return( (long int)bigint );
 
     case INTEGER:
     case CHAR:
